+++ src/command_modules/azure-cli-acr/azure/cli/command_modules/acr/custom.py
@@ -3,28 +3,21 @@
 # Licensed under the MIT License. See License.txt in the project root for license information.
 # --------------------------------------------------------------------------------------------
 
-from subprocess import call, PIPE
-
 from azure.cli.core.commands import LongRunningOperation
 import azure.cli.core.azlogging as azlogging
 from azure.cli.core.util import CLIError
 from azure.cli.core.prompting import prompt, prompt_pass, NoTTYException
 
-from azure.mgmt.containerregistry.v2017_03_01.models import (
-    RegistryUpdateParameters as BasicRegistryUpdateParameters,
-    StorageAccountParameters
-)
-from azure.mgmt.containerregistry.v2017_06_01_preview.models import (
-    RegistryUpdateParameters as ManagedRegistryUpdateParameters,
+from azure.mgmt.containerregistry.v2017_10_01.models import (
+    RegistryUpdateParameters,
+    StorageAccountProperties,
     SkuName,
     SkuTier,
     Sku
 )
 
-from ._constants import MANAGED_REGISTRY_API_VERSION
 from ._factory import get_acr_service_client
 from ._utils import (
-    get_resource_group_name_by_registry_name,
     arm_deploy_template_new_storage,
     arm_deploy_template_existing_storage,
     arm_deploy_template_managed_storage,
@@ -32,7 +25,8 @@ from ._utils import (
     get_registry_by_name,
     managed_registry_validation,
     validate_sku_update,
-    ensure_storage_account_parameter
+    get_resource_group_name_by_registry_name,
+    get_resource_id_by_storage_account_name
 )
 from ._docker_utils import get_login_refresh_token
 from .credential import acr_credential_show
@@ -81,7 +75,11 @@ def acr_create(registry_name,
     client = get_acr_service_client().registries
     admin_user_enabled = admin_enabled == 'true'
 
-    if sku == SkuName.basic.value:
+    if sku == 'Basic':
+        sku = SkuName.classic.value
+        logger.warning("Basic SKU will be deprecated and is converted to Classic SKU.")
+
+    if sku == SkuName.classic.value:
         if storage_account_name is None:
             storage_account_name = random_storage_account_name(registry_name)
             logger.warning(
@@ -161,20 +159,6 @@ def acr_show(registry_name, resource_group_name=None):
     return client.get(resource_group_name, registry_name)
 
 
-def acr_update_get(client,  # pylint: disable=unused-argument
-                   registry_name,
-                   resource_group_name=None):
-    """Returns an empty RegistryUpdateParameters object.
-    :param str registry_name: The name of container registry
-    :param str resource_group_name: The name of resource group
-    """
-    try:
-        managed_registry_validation(registry_name, resource_group_name)
-        return ManagedRegistryUpdateParameters()
-    except:  # pylint: disable=bare-except
-        return BasicRegistryUpdateParameters()
-
-
 def acr_update_custom(instance,
                       sku=None,
                       storage_account_name=None,
@@ -184,7 +168,8 @@ def acr_update_custom(instance,
         instance.sku = Sku(name=sku)
 
     if storage_account_name is not None:
-        instance.storage_account = StorageAccountParameters(storage_account_name, "")
+        instance.storage_account = StorageAccountProperties(
+            get_resource_id_by_storage_account_name(storage_account_name))
 
     if admin_enabled is not None:
         instance.admin_user_enabled = admin_enabled == 'true'
@@ -193,6 +178,12 @@ def acr_update_custom(instance,
         instance.tags = tags
 
     return instance
+
+
+def acr_update_get(client):  # pylint: disable=unused-argument
+    """Returns an empty RegistryUpdateParameters object.
+    """
+    return RegistryUpdateParameters()
 
 
 def acr_update_set(client,
@@ -207,23 +198,37 @@ def acr_update_set(client,
     registry, resource_group_name = get_registry_by_name(registry_name, resource_group_name)
 
     if registry.sku.tier == SkuTier.managed.value:
-        if parameters.sku is not None:
-            validate_sku_update(parameters.sku)
+        validate_sku_update(registry.sku.name, parameters.sku)
         if parameters.storage_account is not None:
             parameters.storage_account = None
             logger.warning(
                 "The registry '%s' in '%s' SKU is a managed registry. The specified storage account will be ignored.",
                 registry_name, registry.sku.name)
-        client = get_acr_service_client(MANAGED_REGISTRY_API_VERSION).registries
-    elif registry.sku.tier == SkuTier.basic.value:
-        if hasattr(parameters, 'sku') and parameters.sku is not None:
-            parameters.sku = None
-            logger.warning(
-                "Updating SKU is not supported for registries in Basic SKU. The specified SKU will be ignored.")
-        if parameters.storage_account is not None:
-            parameters.storage_account = ensure_storage_account_parameter(parameters.storage_account)
+    elif registry.sku.tier == SkuTier.classic.value:
+        validate_sku_update(registry.sku.name, parameters.sku)
 
     return client.update(resource_group_name, registry_name, parameters)
+
+
+def _try_aad_token_login(login_server):
+    try:
+        username = "00000000-0000-0000-0000-000000000000"
+        password = get_login_refresh_token(login_server)
+        return username, password
+    except CLIError as e:
+        logger.warning("AAD authentication failed with message: %s", str(e))
+        return None, None
+
+
+def _try_admin_user_login(registry_name):
+    try:
+        cred = acr_credential_show(registry_name)
+        username = cred.username
+        password = cred.passwords[0].value
+        return username, password
+    except CLIError as e:
+        logger.warning("Admin user authentication failed with message: %s", str(e))
+        return None, None
 
 
 def acr_login(registry_name, resource_group_name=None, username=None, password=None):
@@ -233,10 +238,20 @@ def acr_login(registry_name, resource_group_name=None, username=None, password=N
     :param str username: The username used to log into the container registry
     :param str password: The password used to log into the container registry
     """
+    from subprocess import PIPE, Popen, SubprocessError
+    docker_not_installed = "Please verify if docker is installed."
+    docker_not_available = "Please verify if docker daemon is running properly."
+
     try:
-        call(["docker", "ps"], stdout=PIPE, stderr=PIPE)
-    except:
-        raise CLIError("Please verify whether docker is installed and running properly")
+        p = Popen(["docker", "ps"], stdout=PIPE, stderr=PIPE)
+        returncode = p.wait()
+    except OSError:
+        raise CLIError(docker_not_installed)
+    except SubprocessError:
+        raise CLIError(docker_not_available)
+
+    if returncode:
+        raise CLIError(docker_not_available)
 
     registry, _ = get_registry_by_name(registry_name, resource_group_name)
     sku_tier = registry.sku.tier
@@ -249,23 +264,13 @@ def acr_login(registry_name, resource_group_name=None, username=None, password=N
         except NoTTYException:
             raise CLIError('Please specify both username and password in non-interactive mode.')
 
-    if sku_tier == SkuTier.managed.value:
-        # 2. if we don't yet have credentials, attempt to get a refresh token
-        if not password:
-            try:
-                username = "00000000-0000-0000-0000-000000000000"
-                password = get_login_refresh_token(login_server)
-            except Exception as e:  # pylint: disable=broad-except
-                logger.warning("AAD authentication failed with message: %s", str(e))
+    # 2. if we don't yet have credentials, attempt to get a refresh token
+    if not password and sku_tier == SkuTier.managed.value:
+        username, password = _try_aad_token_login(login_server)
 
     # 3. if we still don't have credentials, attempt to get the admin credentials (if enabled)
     if not password:
-        try:
-            cred = acr_credential_show(registry_name)
-            username = cred.username
-            password = cred.passwords[0].value
-        except Exception as e:  # pylint: disable=broad-except
-            logger.warning("Admin user authentication failed with message: %s", str(e))
+        username, password = _try_admin_user_login(registry_name)
 
     # 4. if we still don't have credentials, prompt the user
     if not password:
@@ -277,10 +282,29 @@ def acr_login(registry_name, resource_group_name=None, username=None, password=N
                 'Unable to authenticate using AAD or admin login credentials. ' +
                 'Please specify both username and password in non-interactive mode.')
 
-    call(["docker", "login",
-          "--username", username,
-          "--password", password,
-          login_server])
+    use_password_stdin = False
+    try:
+        p = Popen(["docker", "login", "--help"], stdout=PIPE, stderr=PIPE)
+        stdout, _ = p.communicate()
+        if b'--password-stdin' in stdout:
+            use_password_stdin = True
+    except OSError:
+        raise CLIError(docker_not_installed)
+    except SubprocessError:
+        raise CLIError(docker_not_available)
+
+    if use_password_stdin:
+        p = Popen(["docker", "login",
+                   "--username", username,
+                   "--password-stdin",
+                   login_server], stdin=PIPE)
+        p.communicate(input=password.encode())
+    else:
+        p = Popen(["docker", "login",
+                   "--username", username,
+                   "--password", password,
+                   login_server])
+        p.wait()
 
 
 def acr_show_usage(registry_name, resource_group_name=None):
@@ -289,7 +313,7 @@ def acr_show_usage(registry_name, resource_group_name=None):
     :param str resource_group_name: The name of resource group
     """
     _, resource_group_name = managed_registry_validation(
-        registry_name, resource_group_name, "Usage is not supported for registries in Basic SKU.")
-    client = get_acr_service_client(MANAGED_REGISTRY_API_VERSION).registries
+        registry_name, resource_group_name, "Usage is only supported for managed registries.")
+    client = get_acr_service_client().registries
 
     return client.list_usages(resource_group_name, registry_name)